+++ src/App.js
@@ -4,18 +4,24 @@ import Dashboard from './components/Dashboard/Dashboard';
 import { Route, BrowserRouter as Router, Switch } from 'react-router-dom';
 import PrivateRoute from './components/PrivateRoute';
 import { Profile } from './components/Profile'
-import Login from './Pages/Login';
+import Login from './components/UserPathLog/Login';
 import FindPeople from './components/Friends/FindPeople'
-import OtherProfile from './components/Friends/OtherProfile'
+import OtherProfile from './components/Friends/OtherProfile';
+import Registration from './components/UserPathLog/Register';
+import Resetpw from './components/UserPathLog/Resetpw';
+import { HashRouter } from "react-router-dom";
+
 // import Profile from './components/Profile'
-import { TestLogin } from './Pages/TestLogin';
-import { CirclesBackground } from "./components/styleElements/CirclesBackground";
-import { BottomNav } from "./components/BottomNav";
+// import { TestLogin } from './components/TestLogin/TestLogin';
+// import { CirclesBackground } from "./components/styleElements/CirclesBackground";
+// import { BottomNav } from "./components/BottomNav";
+
 
 const App = () => {
 	return (
 		<Router>
 			<Switch>
+				<a className="logoref" href="http://localhost:3000/welcome#/"> </a>
 				<PrivateRoute path='/dashboard'>
 					<Dashboard />
 				</PrivateRoute>
@@ -34,14 +40,17 @@ const App = () => {
 				<Route path="/findpeople">
 					<FindPeople />
 				</Route>
-
-				<Route path="/">
+				{/* <HashRouter> */}
+				{/* <Route path="/" >
 					<Login />
-				</Route>
-
-				<Route path="/user/:id" component={OtherProfile}></Route>
+				</Route> */}
+				{/* <Route exact path="/register" component={Registration}> Register </Route>
+				<Route exact path="/resetpw" component={Resetpw}> Reset PW</Route> */}
+				{/* </HashRouter> */}
+				{/* add Register and Reset PW */}
+				{/* <Route path="/user/:id" component={OtherProfile}></Route> */}
 			</Switch>
-		</Router>
+		</Router >
 	);
 };
