--- src/Pages/Movies.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function Movies() {
-	return <p> Movies</p>;
-}
