+++ src/components/UserPathLog/Login.js
@@ -1,6 +1,10 @@
 import Axios from "axios";
 import React, { useState } from "react";
 import { useHistory } from "react-router-dom";
+import Registration from '../UserPathLog/Register';
+import Resetpw from '../UserPathLog/Resetpw';
+import { HashRouter, Route } from "react-router-dom";
+
 
 export default function Login() {
 	const [username, setUsername] = useState("");
@@ -32,22 +36,27 @@ export default function Login() {
 	}
 
 	return (
-		<div className="login">
-			<h1>Login Page</h1>
-			<label> user name </label>
-			<input
-				type="text"
-				placeholder="username"
-				onChange={(e) => setUsername(e.target.value)}
-				value={username}
-			/>
-			<label> password </label>
-			<input
-				type="password"
-				onChange={(e) => setPassword(e.target.value)}
-				value={password}
-			/>
-			<button onClick={login}> Login </button>
-		</div >
+		<React.Fragment>
+			<div className="login">
+				<h1>Login Page</h1>
+				<label> user name </label>
+				<input
+					type="text"
+					placeholder="username"
+					onChange={(e) => setUsername(e.target.value)}
+					value={username}
+				/>
+				<label> password </label>
+				<input
+					type="password"
+					onChange={(e) => setPassword(e.target.value)}
+					value={password}
+				/>
+				<button onClick={login}> Login </button>
+			</div >
+
+
+
+		</React.Fragment>
 	);
 }
